requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.8
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   openai
arxiv==2.1.3
    # via -r requirements.in
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backports-tarfile==1.2.0
    # via jaraco-context
beautifulsoup4==4.12.3
    # via wikipedia
blinker==1.9.0
    # via streamlit
brotli==1.1.0
    # via httpx
cachetools==5.5.0
    # via
    #   google-auth
    #   streamlit
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   streamlit
config==0.5.1
    # via -r requirements.in
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
duckduckgo-search==5.3.1b1
    # via -r requirements.in
feedparser==6.0.11
    # via arxiv
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-api-core==2.23.0
    # via google-api-python-client
google-api-python-client==2.154.0
    # via -r requirements.in
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.66.0
    # via google-api-core
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx[brotli,http2,socks]==0.27.2
    # via
    #   duckduckgo-search
    #   langsmith
    #   openai
    #   wolframalpha
httpx-sse==0.4.0
    # via langchain-community
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jaraco-context==6.0.1
    # via wolframalpha
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jiter==0.8.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.9
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.8
    # via
    #   -r requirements.in
    #   langchain-experimental
langchain-core==0.3.21
    # via
    #   langchain
    #   langchain-community
    #   langchain-experimental
    #   langchain-openai
    #   langchain-text-splitters
langchain-experimental==0.3.3
    # via -r requirements.in
langchain-openai==0.2.10
    # via -r requirements.in
langchain-text-splitters==0.3.2
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.23.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via wolframalpha
multidict==6.1.0
    # via
    #   aiohttp
    #   wolframalpha
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.14.2
    # via altair
numexpr==2.10.2
    # via -r requirements.in
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   numexpr
    #   pandas
    #   pydeck
    #   streamlit
openai==1.55.2
    # via langchain-openai
orjson==3.10.12
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.0.0
    # via streamlit
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.25.0
    # via google-api-core
protobuf==5.29.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==18.1.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   arxiv
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
    #   wikipedia
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
socksio==1.0.0
    # via httpx
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.35
    # via
    #   langchain
    #   langchain-community
streamlit==1.40.2
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.8.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   altair
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
wikipedia==1.4.0
    # via -r requirements.in
wolframalpha==5.1.3
    # via -r requirements.in
xmltodict==0.14.2
    # via wolframalpha
yarl==1.18.0
    # via aiohttp

langchain_google_genai